app/registries/Registry.ts
import axios, {
    type AxiosRequestConfig,
    type Method,
    type AxiosResponse,
} from 'axios';
import log from '../log/index.js';
import Component from '../registry/Component.js';
import { getSummaryTags } from '../prometheus/registry.js';
import type { ContainerImage } from '../model/container.js';

export interface RegistryImage extends ContainerImage {
    // Add any registry specific properties if needed
}

export interface RegistryManifest {
    digest?: string;
    version?: number;
    created?: string;
}

export interface RegistryTagsList {
    name: string;
    tags: string[];
}

export interface RegistryManifestResponse {
    schemaVersion: number;
    mediaType?: string;
    manifests?: {
        digest: string;
        mediaType: string;
        platform: {
            architecture: string;
            os: string;
            variant?: string;
        };
    }[];
    config?: {
        digest: string;
        mediaType: string;
    };
    history?: {
        v1Compatibility: string;
    }[];
}

/**
 * Docker Registry Abstract class.
 */
class Registry extends Component {
    /**
     * Encode Bse64(login:password)
     * @param login
     * @param token
     * @returns {string}
     */
    static base64Encode(login: string, token: string) {
        return Buffer.from(`${login}:${token}`, 'utf-8').toString('base64');
    }

    /**
     * If this registry is responsible for the image (to be overridden).
     * @param image the image
     * @returns {boolean}
     */
    match(_image: ContainerImage): boolean {
        return false;
    }

    /**
     * Normalize image according to Registry Custom characteristics (to be overridden).
     * @param image
     * @returns {*}
     */
    normalizeImage(image: ContainerImage): ContainerImage {
        return image;
    }

    /**
     * Authenticate and set authentication value to requestOptions.
     * @param image
     * @param requestOptions
     * @returns {*}
     */
    async authenticate(
        _image: ContainerImage,
        requestOptions: AxiosRequestConfig,
    ): Promise<AxiosRequestConfig> {
        return requestOptions;
    }

    /**
     * Get Tags.
     * @param image
     * @returns {*}
     */
    async getTags(image: ContainerImage): Promise<string[]> {
        this.log.debug(`Get ${image.name} tags`);
        const tags: string[] = [];
        let page: AxiosResponse<RegistryTagsList> | undefined = undefined;
        let hasNext = true;
        let link: string | undefined = undefined;
        while (hasNext) {
            const lastItem =
                page && page.data && page.data.tags
                    ? page.data.tags[page.data.tags.length - 1]
                    : undefined;

            page = await this.getTagsPage(image, lastItem, link);
            const pageTags =
                page && page.data && page.data.tags ? page.data.tags : [];
            link = page && page.headers ? page.headers.link : undefined;
            hasNext = page && page.headers && page.headers.link !== undefined;
            tags.push(...pageTags);
        }

        // Sort alpha then reverse to get higher values first
        tags.sort();
        tags.reverse();
        return tags;
    }

    /**
     * Get tags page
     * @param image
     * @param lastItem
     * @returns {Promise<*>}
     */
    getTagsPage(
        image: ContainerImage,
        lastItem: string | undefined = undefined,
        _link: string | undefined = undefined,
    ) {
        // Default items per page (not honoured by all registries)
        const itemsPerPage = 1000;
        const last = lastItem ? `&last=${lastItem}` : '';
        return this.callRegistry<RegistryTagsList>({
            image,
            url: `${image.registry.url}/${image.name}/tags/list?n=${itemsPerPage}${last}`,
            resolveWithFullResponse: true,
        });
    }

    /**
     * Get image manifest for a remote tag.
     * @param image
     * @param digest (optional)
     * @returns {Promise<undefined|*>}
     */
    async getImageManifestDigest(
        image: ContainerImage,
        digest?: string,
    ): Promise<RegistryManifest> {
        const tagOrDigest = digest || image.tag.value;
        let manifestDigestFound;
        let manifestMediaType;
        this.log.debug(
            `${this.getId()} - Get ${image.name}:${tagOrDigest} manifest`,
        );
        const responseManifests =
            await this.callRegistry<RegistryManifestResponse>({
                image,
                url: `${image.registry.url}/${image.name}/manifests/${tagOrDigest}`,
                headers: {
                    Accept: 'application/vnd.docker.distribution.manifest.list.v2+json, application/vnd.oci.image.index.v1+json, application/vnd.docker.distribution.manifest.v2+json, application/vnd.oci.image.manifest.v1+json',
                },
            });
        if (responseManifests) {
            log.debug(`Found manifests [${JSON.stringify(responseManifests)}]`);
            if (responseManifests.schemaVersion === 2) {
                log.debug('Manifests found with schemaVersion = 2');
                log.debug(
                    `Manifests media type detected [${responseManifests.mediaType}]`,
                );
                if (
                    responseManifests.mediaType ===
                        'application/vnd.docker.distribution.manifest.list.v2+json' ||
                    responseManifests.mediaType ===
                        'application/vnd.oci.image.index.v1+json'
                ) {
                    log.debug(
                        `Filter manifest for [arch=${image.architecture}, os=${image.os}, variant=${image.variant}]`,
                    );
                    let manifestFound;
                    const manifestFounds = responseManifests.manifests.filter(
                        (manifest: any) =>
                            manifest.platform.architecture ===
                                image.architecture &&
                            manifest.platform.os === image.os,
                    );

                    // 1 manifest matching al least? Get the first one (better than nothing)
                    if (manifestFounds.length > 0) {
                        [manifestFound] = manifestFounds;
                    }

                    // Multiple matching manifests? Try to refine using variant filtering
                    if (manifestFounds.length > 1) {
                        const manifestFoundFilteredOnVariant =
                            manifestFounds.find(
                                (manifest: any) =>
                                    manifest.platform.variant === image.variant,
                            );

                        // Manifest exactly matching with variant? Select it
                        if (manifestFoundFilteredOnVariant) {
                            manifestFound = manifestFoundFilteredOnVariant;
                        }
                    }

                    if (manifestFound) {
                        log.debug(
                            `Manifest found with [digest=${manifestFound.digest}, mediaType=${manifestFound.mediaType}]`,
                        );
                        manifestDigestFound = manifestFound.digest;
                        manifestMediaType = manifestFound.mediaType;
                    }
                } else if (
                    responseManifests.mediaType ===
                        'application/vnd.docker.distribution.manifest.v2+json' ||
                    responseManifests.mediaType ===
                        'application/vnd.oci.image.manifest.v1+json'
                ) {
                    const manifestReference = manifestDigestFound || tagOrDigest;
                    log.debug(
                        `Manifest found with [reference=${manifestReference}, mediaType=${responseManifests.mediaType}]`,
                    );
                    manifestDigestFound = manifestReference;
                    manifestMediaType = responseManifests.mediaType;
                }
            } else if (responseManifests.schemaVersion === 1) {
                log.debug('Manifests found with schemaVersion = 1');
                const v1Compat = JSON.parse(
                    responseManifests.history[0].v1Compatibility,
                );
                const manifestFound = {
                    digest: v1Compat.config ? v1Compat.config.Image : undefined,
                    created: v1Compat.created,
                    version: 1,
                };
                log.debug(
                    `Manifest found with [digest=${manifestFound.digest}, created=${manifestFound.created}, version=${manifestFound.version}]`,
                );
                return manifestFound;
            }
            if (
                (manifestDigestFound &&
                    manifestMediaType ===
                        'application/vnd.docker.distribution.manifest.v2+json') ||
                (manifestDigestFound &&
                    manifestMediaType ===
                        'application/vnd.oci.image.manifest.v1+json')
            ) {
                log.debug(
                    'Calling registry to get docker-content-digest header',
                );
                const responseManifest =
                    await this.callRegistry<RegistryManifestResponse>({
                        image,
                        method: 'head',
                        url: `${image.registry.url}/${image.name}/manifests/${manifestDigestFound}`,
                        headers: {
                            Accept: manifestMediaType,
                        },
                        resolveWithFullResponse: true,
                    });
                const manifestFound = {
                    digest: responseManifest.headers['docker-content-digest'],
                    version: 2,
                };
                log.debug(
                    `Manifest found with [digest=${manifestFound.digest}, version=${manifestFound.version}]`,
                );
                return manifestFound;
            }
            if (
                (manifestDigestFound &&
                    manifestMediaType ===
                        'application/vnd.docker.container.image.v1+json') ||
                (manifestDigestFound &&
                    manifestMediaType ===
                        'application/vnd.oci.image.config.v1+json')
            ) {
                const manifestFound = {
                    digest: manifestDigestFound,
                    version: 1,
                };
                log.debug(
                    `Manifest found with [digest=${manifestFound.digest}, version=${manifestFound.version}]`,
                );
                return manifestFound;
            }
        }
        // Empty result...
        throw new Error('Unexpected error; no manifest found');
    }

    async callRegistry<T = any>(options: {
        image: ContainerImage;
        url: string;
        method?: Method;
        headers?: any;
        resolveWithFullResponse: true;
    }): Promise<AxiosResponse<T>>;

    async callRegistry<T = any>(options: {
        image: ContainerImage;
        url: string;
        method?: Method;
        headers?: any;
        resolveWithFullResponse?: false;
    }): Promise<T>;

    async callRegistry<T = any>({
        image,
        url,
        method = 'get',
        headers = {
            Accept: 'application/json',
        },
        resolveWithFullResponse = false,
    }: {
        image: ContainerImage;
        url: string;
        method?: Method;
        headers?: any;
        resolveWithFullResponse?: boolean;
    }): Promise<T | AxiosResponse<T>> {
        const start = Date.now();

        // Request options
        const axiosOptions: AxiosRequestConfig = {
            url,
            method,
            headers,
            responseType: 'json',
        };

        const axiosOptionsWithAuth = await this.authenticate(
            image,
            axiosOptions,
        );

        try {
            const response = (await axios(
                axiosOptionsWithAuth,
            )) as AxiosResponse<T>;
            const end = Date.now();
            getSummaryTags().observe(
                { type: this.type, name: this.name },
                (end - start) / 1000,
            );
            return resolveWithFullResponse ? response : response.data;
        } catch (error) {
            const end = Date.now();
            getSummaryTags().observe(
                { type: this.type, name: this.name },
                (end - start) / 1000,
            );
            throw error;
        }
    }

    getImageFullName(image: ContainerImage, tagOrDigest: string) {
        // digests are separated with @ whereas tags are separated with :
        const tagOrDigestWithSeparator =
            tagOrDigest.indexOf(':') !== -1
                ? `@${tagOrDigest}`
                : `:${tagOrDigest}`;
        let fullName = `${image.registry.url}/${image.name}${tagOrDigestWithSeparator}`;

        fullName = fullName.replace(/https?:\/\//, '');
        fullName = fullName.replace(/\/v2/, '');
        return fullName;
    }

    /**
     * Return {username, pass } or undefined.
     * @returns {}
     */

    async getAuthPull(): Promise<{ username?: string; password?: string } | undefined> {
        return undefined;
    }
}

export default Registry;
